# Speed up compilation