--- a/devito/core/operator.py
+++ b/devito/core/operator.py
@@ -9,13 +9,6 @@ __all__ = ['OperatorCore']
 
 class OperatorCore(Operator):
 
-    @classmethod
-    def _specialize_exprs(cls, expressions):
-        # Align data accesses to the computational domain
-        key = lambda i: i.is_DiscreteFunction
-        expressions = [align_accesses(e, key=key) for e in expressions]
-        return super(OperatorCore, cls)._specialize_exprs(expressions)
-
     def _autotune(self, args, setup):
         if setup in [False, 'off']:
             return args