--- a/devito/finite_differences/differentiable.py
+++ b/devito/finite_differences/differentiable.py
@@ -6,8 +6,9 @@ from sympy.functions.elementary.integers import floor
 from sympy.core.evalf import evalf_table
 
 from cached_property import cached_property
+from devito.finite_differences.lazy import Evaluable
 from devito.logger import warning
-from devito.tools import Evaluable, EnrichedTuple, filter_ordered, flatten
+from devito.tools import EnrichedTuple, filter_ordered, flatten
 
 __all__ = ['Differentiable']