… devito/ir/iet/visitors.py
… devito/ir/support/space.py
… devito/ir/support/utils.py
… devito/ir/support/vector.py
… devito/operations/interpolators.py
… devito/operator/operator.py
… devito/passes/clusters/aliases.py
… devito/passes/clusters/blocking.py
… devito/passes/clusters/cse.py
… devito/passes/clusters/factorization.py
… devito/passes/clusters/misc.py
… devito/passes/clusters/utils.py
… devito/symbolics/extended_sympy.py
… devito/symbolics/manipulation.py
… devito/symbolics/printer.py
… devito/symbolics/queries.py
… devito/tools/abc.py
… devito/types/basic.py
… devito/types/dense.py
… devito/types/dimension.py
… devito/types/equation.py
… devito/types/sparse.py
… devito/types/tensor.py
… tests/test_dse.py
… tests/test_operator.py