A collection of self-contained, orthogonal commits. I could have filed 20 PRs, one per commit, or just one.

On my workstation, the compilation times w.r.t. `master` become
- tti so12 : ~27s -> ~12s
- smoothers : ~5.5s -> ~1s
- elastic so12: ~18s  -> ~8.5s
- viscoelastic so12 : ~41s -> ~14s
- iso-acoustic so12 :  ~4s -> ~2.5s

@tjb900 I'm not sure if you care about this ... tagging you, just in case

@ggorman I've been working on this intermittently for a few days, but there is still work to do

@mloubout @rhodrin if you take a profile with Python's cProfile, you shall see there's a margin of improvement especially in the DSL/SymPy layers... wonder whether/when you could this give a go. Not a top priority of course!