@georgebisbas 2020-04-02
On my Azure workstation, the compilation times w.r.t. master become with Intel Compiler:
(DEVITO_ARCH=icc, DEVITO_OPENMP=1, 10 threads-ht off)
    tti so12 : ~5s -> ~5s
    elastic so12: ~12s -> ~12s
    viscoelastic so12 : ~16s -> ~14s
    iso-acoustic so12 : ~2s -> ~2s

Am I doing sth wrong?